auth/requirements.txt
alembic==1.16.4
annotated-types==0.7.0
anyio==4.10.0
astroid==3.3.11
bcrypt==4.3.0
click==8.1.8
dill==0.4.0
ecdsa==0.19.1
exceptiongroup==1.3.0
fastapi==0.116.1
h11==0.16.0
idna==3.10
isort==6.0.1
jedi==0.19.2
Mako==1.3.10
MarkupSafe==3.0.2
mccabe==0.7.0
mysqlclient==2.2.7
parso==0.8.4
passlib==1.7.4
platformdirs==4.3.8
pyasn1==0.6.1
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
PyJWT==2.10.1
pylint==3.3.8
python-dotenv==1.1.1
python-jose==3.5.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.43
starlette==0.47.2
tomli==2.2.1
tomlkit==0.13.3
typing-inspection==0.4.1
typing_extensions==4.14.1
uvicorn==0.35.0
